refactor(storage): extract RunToolDependencyStore sub-mixin (SQLite + ABC)
Final bucket of the _agent_run.py decomposition. Move the four
run-tool-dependency methods (attach_run_tool_dependency,
count_unresolved_followup_dependencies, list_run_tool_dependencies,
consume_run_tool_dependencies) verbatim into RunToolDependencyStoreABC and
SQLiteRunToolDependencyStoreMixin. Preserves the consume resolved-unconsumed
WHERE guard + rowcount and the count three-table JOIN byte-identically.

The composite AgentRunMixin now inherits all three agent-run ABCs
(AgentRunStoreABC, PendingToolCallStoreABC, RunToolDependencyStoreABC). The
drained concrete residual SQLiteAgentRunMixin class is removed from the storage
base list; the residual _agent_run.py module stays as the helpers-only home of
the shared _dt/_row_to_* helpers imported by all three sub-mixins. All 23
agent-run methods now live in the three sub-mixins.
