+++ reflexio/server/services/storage/sqlite_storage/__init__.py
@@ -1,4 +1,3 @@
-from ._agent_run import SQLiteAgentRunMixin
 from ._base import (
     SQLiteStorageBase,
     _cosine_similarity,
@@ -24,7 +23,11 @@ from ._stall_state import (
     mark_stall_notified,
     upsert_stall_state,
 )
-from .agent_run import SQLiteAgentRunStoreMixin, SQLitePendingToolCallStoreMixin
+from .agent_run import (
+    SQLiteAgentRunStoreMixin,
+    SQLitePendingToolCallStoreMixin,
+    SQLiteRunToolDependencyStoreMixin,
+)
 from .base import SQLiteDeletionMixin, SQLiteFtsVecMixin
 from .governance import (
     AuditEventStoreMixin,
@@ -46,7 +49,7 @@ from .profiles import InteractionStoreMixin, ProfileSearchMixin, ProfileStoreMix
 class SQLiteStorage(
     SQLiteAgentRunStoreMixin,
     SQLitePendingToolCallStoreMixin,
-    SQLiteAgentRunMixin,
+    SQLiteRunToolDependencyStoreMixin,
     ProfileStoreMixin,
     InteractionStoreMixin,
     ProfileSearchMixin,
