+++ reflexio/server/services/storage/sqlite_storage/_agent_run.py
@@ -1,10 +1,16 @@
-"""SQLite storage for resumable extraction agent runs."""
+"""Shared row/datetime helpers for SQLite resumable extraction agent runs.
+
+Helpers-only residual for the ``_agent_run.py`` decomposition: all twenty-three
+agent-run methods now live in the three ``agent_run`` sub-mixins
+(``SQLiteAgentRunStoreMixin`` / ``SQLitePendingToolCallStoreMixin`` /
+``SQLiteRunToolDependencyStoreMixin``), which import the ``_row_to_*`` /
+``_dt`` / ``_dt_str`` helpers below rather than duplicating them.
+"""
 
 from __future__ import annotations
 
 import sqlite3
 from datetime import UTC, datetime
-from typing import Any
 
 from reflexio.server.services.storage.storage_base import (
     AgentBinding,
@@ -18,7 +24,7 @@ from reflexio.server.services.storage.storage_base import (
     embedding_similarity,
 )
 
-from ._base import SQLiteStorageBase, _json_loads
+from ._base import _json_loads
 
 
 def _dt(value: str | None) -> datetime | None:
@@ -136,101 +142,3 @@ def _row_to_run_tool_dependency(row: sqlite3.Row) -> RunToolDependencyRecord:
         consumed_at=_dt(data.get("consumed_at")),
         created_at=_dt(data.get("created_at")),
     )
-
-
-class SQLiteAgentRunMixin:
-    """SQLite-backed resumable extraction run storage."""
-
-    _lock: Any
-    conn: sqlite3.Connection
-    _fetchone: Any
-    _fetchall: Any
-    _current_timestamp: Any
-    org_id: str
-
-    @SQLiteStorageBase.handle_exceptions
-    def attach_run_tool_dependency(
-        self, record: RunToolDependencyRecord
-    ) -> RunToolDependencyRecord:
-        with self._lock:
-            self.conn.execute(
-                """
-                INSERT OR IGNORE INTO _run_tool_dependencies (
-                    run_id, pending_tool_call_id, dependency_kind, resolved_at,
-                    consumed_at
-                ) VALUES (?,?,?,?,?)
-                """,
-                (
-                    record.run_id,
-                    record.pending_tool_call_id,
-                    record.dependency_kind.value,
-                    _dt_str(record.resolved_at),
-                    _dt_str(record.consumed_at),
-                ),
-            )
-            self.conn.commit()
-        row = self._fetchone(
-            """
-            SELECT * FROM _run_tool_dependencies
-            WHERE run_id = ? AND pending_tool_call_id = ?
-            """,
-            (record.run_id, record.pending_tool_call_id),
-        )
-        if row is None:  # pragma: no cover
-            raise RuntimeError("Failed to attach run tool dependency")
-        return _row_to_run_tool_dependency(row)
-
-    @SQLiteStorageBase.handle_exceptions
-    def count_unresolved_followup_dependencies(
-        self,
-        *,
-        org_id: str,
-        extractor_kind: str,
-        tool_name: str,
-    ) -> int:
-        row = self._fetchone(
-            """
-            SELECT COUNT(*) AS count
-            FROM _run_tool_dependencies d
-            JOIN _agent_runs r ON r.id = d.run_id
-            JOIN _pending_tool_calls p ON p.id = d.pending_tool_call_id
-            WHERE r.org_id = ?
-              AND r.extractor_kind = ?
-              AND p.tool_name = ?
-              AND p.status = ?
-              AND d.resolved_at IS NULL
-              AND d.consumed_at IS NULL
-            """,
-            (
-                org_id,
-                extractor_kind,
-                tool_name,
-                PendingToolCallStatus.PENDING.value,
-            ),
-        )
-        return int(row["count"]) if row is not None else 0
-
-    @SQLiteStorageBase.handle_exceptions
-    def list_run_tool_dependencies(self, run_id: str) -> list[RunToolDependencyRecord]:
-        rows = self._fetchall(
-            "SELECT * FROM _run_tool_dependencies WHERE run_id = ?",
-            (run_id,),
-        )
-        return [_row_to_run_tool_dependency(row) for row in rows]
-
-    @SQLiteStorageBase.handle_exceptions
-    def consume_run_tool_dependencies(self, run_id: str) -> int:
-        consumed_at = self._current_timestamp()
-        with self._lock:
-            cur = self.conn.execute(
-                """
-                UPDATE _run_tool_dependencies
-                SET consumed_at = ?
-                WHERE run_id = ?
-                  AND resolved_at IS NOT NULL
-                  AND consumed_at IS NULL
-                """,
-                (consumed_at, run_id),
-            )
-            self.conn.commit()
-        return int(cur.rowcount)
