+++ reflexio/server/services/storage/sqlite_storage/agent_run/__init__.py
@@ -1,4 +1,9 @@
 from ._agent_run_store import SQLiteAgentRunStoreMixin
 from ._pending_tool_call_store import SQLitePendingToolCallStoreMixin
+from ._run_tool_dependency_store import SQLiteRunToolDependencyStoreMixin
 
-__all__ = ["SQLiteAgentRunStoreMixin", "SQLitePendingToolCallStoreMixin"]
+__all__ = [
+    "SQLiteAgentRunStoreMixin",
+    "SQLitePendingToolCallStoreMixin",
+    "SQLiteRunToolDependencyStoreMixin",
+]
