+++ reflexio/server/services/storage/storage_base/agent_run/__init__.py
@@ -1,4 +1,9 @@
 from ._agent_run_store import AgentRunStoreABC
 from ._pending_tool_call_store import PendingToolCallStoreABC
+from ._run_tool_dependency_store import RunToolDependencyStoreABC
 
-__all__ = ["AgentRunStoreABC", "PendingToolCallStoreABC"]
+__all__ = [
+    "AgentRunStoreABC",
+    "PendingToolCallStoreABC",
+    "RunToolDependencyStoreABC",
+]
